chore: create tabs utils
